establish detail page and implement navigation

+++ STYLiSH/app/src/main/java/com/yazhiyue/stylish/data/Variant.kt
@@ -5,5 +5,5 @@ import com.squareup.moshi.Json
 data class Variant(
     @Json(name = "color_code") val colorCode: String,
     val size: String,
-    val stock: Long
+    val stock: Int
 )
